+++ jgiven-core/src/main/java/com/tngtech/jgiven/annotation/IsTag.java
@@ -1,5 +1,7 @@
 package com.tngtech.jgiven.annotation;
 
+import com.tngtech.jgiven.impl.tag.DefaultTagDescriptionGenerator;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;